YOLOv8/requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy<2.0.0
torch>=1.11.0
torchvision>=0.12.0
